+++ package.json
@@ -13,6 +13,7 @@
     "@babel/core": "7.6.4",
     "@babel/preset-env": "7.6.3",
     "@babel/preset-react": "7.6.3",
+    "axios-mock-adapter": "1.17.0",
     "babel-jest": "24.9.0",
     "babel-loader": "8.0.6",
     "enzyme": "3.10.0",
